move loaded files to new directory
